refactor: validate entries + simplify vocab clear
