+++ invenio_rdm_records/vocabularies/vocabulary.py
@@ -9,11 +9,43 @@
 """Vocabulary."""
 
 import csv
-from collections import OrderedDict
+from collections import OrderedDict, defaultdict
 
 from flask_babelex import lazy_gettext as _
 
-from .utils import hierarchized_rows
+
+def hierarchized_rows(dict_reader):
+    """Yields filled OrderedDict rows according to csv hierarchy.
+
+    Idea is to have the csv rows:
+
+    fooA, barA-1, bazA-1
+        , barA-2, bazA-2
+    fooB, barB-1, bazB-1
+        ,       , bazB-2
+
+    map to these rows
+
+    fooA, barA-1, bazA-1
+    fooA, barA-2, bazA-2
+    fooB, barB-1, bazB-1
+    fooB, barB-1, bazB-2
+
+    This makes it easy for subject matter experts to fill the csv in
+    their spreadsheet software, while also allowing hierarchy of data
+    a-la yaml and extensibility for other conversions or data down the road.
+    """
+    prev_row = defaultdict(lambda: "")
+
+    for row in dict_reader:  # row is an OrderedDict in fieldnames order
+        current_row = row
+        for field in row:
+            if not current_row[field]:
+                current_row[field] = prev_row[field]
+            else:
+                break
+        prev_row = current_row
+        yield current_row
 
 
 class Vocabulary(object):
